Clean up entities
