+++ src/app/core/services/cluster.ts
@@ -27,6 +27,7 @@ import {
   Cluster,
   ClusterPatch,
   CNIPluginVersions,
+  CreateClusterModel,
   Finalizer,
   MasterVersion,
   ProviderSettingsPatch,
@@ -36,7 +37,6 @@ import {Health} from '@shared/entity/health';
 import {ClusterMetrics, NodeMetrics} from '@shared/entity/metrics';
 import {Node} from '@shared/entity/node';
 import {SSHKey} from '@shared/entity/ssh-key';
-import {CreateClusterModel} from '@shared/model/CreateClusterModel';
 import {merge, Observable, of, Subject, timer} from 'rxjs';
 import {catchError, filter, shareReplay, switchMap, switchMapTo, take} from 'rxjs/operators';
 import {ExternalCluster, ExternalClusterModel, ExternalClusterPatch} from '@shared/entity/external-cluster';
