+++ src/app/shared/components/add-project-dialog/component.ts
@@ -18,8 +18,8 @@ import {MatDialogRef} from '@angular/material/dialog';
 import {ApiService} from '@core/services/api';
 import {NotificationService} from '@core/services/notification';
 import {ResourceType} from '@shared/entity/common';
-import {CreateProjectModel} from '@shared/model/CreateProjectModel';
 import {AsyncValidators} from '@shared/validators/async-label-form.validator';
+import {ProjectModel} from '@shared/entity/project';
 
 @Component({
   selector: 'km-add-project-dialog',
@@ -48,7 +48,7 @@ export class AddProjectDialogComponent implements OnInit {
       return;
     }
 
-    const createProject: CreateProjectModel = {
+    const createProject: ProjectModel = {
       name: this.form.controls.name.value,
       labels: this.labels,
     };
