--- src/app/shared/model/CreateClusterModel.ts
@@ -1,30 +1,0 @@
-// Copyright 2020 The Kubermatic Kubernetes Platform contributors.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-import {ClusterSpec} from '../entity/cluster';
-import {MachineDeployment} from '../entity/machine-deployment';
-
-export class CreateClusterModel {
-  cluster: ClusterModel;
-  // NOTE: The Kubermatic API abstraction for MachineDeployments is NodeDeployments
-  nodeDeployment?: MachineDeployment;
-}
-
-class ClusterModel {
-  name: string;
-  spec: ClusterSpec;
-  labels?: object;
-  type: string;
-  credential?: string;
-}
