+++ src/app/wizard/component.ts
@@ -23,10 +23,9 @@ import {NodeDataService} from '@core/services/node-data/service';
 import {NotificationService} from '@core/services/notification';
 import {ProjectService} from '@core/services/project';
 import {WizardService} from '@core/services/wizard/wizard';
-import {Cluster} from '@shared/entity/cluster';
+import {Cluster, CreateClusterModel} from '@shared/entity/cluster';
 import {Project} from '@shared/entity/project';
 import {SSHKey} from '@shared/entity/ssh-key';
-import {CreateClusterModel} from '@shared/model/CreateClusterModel';
 import {NodeData} from '@shared/model/NodeSpecChange';
 import {forkJoin, of, Subject, take} from 'rxjs';
 import {filter, switchMap, takeUntil, tap} from 'rxjs/operators';
